+++ src/main/java/com/koddy/server/member/application/usecase/command/CompleteMentorCommand.java
@@ -1,6 +1,5 @@
 package com.koddy.server.member.application.usecase.command;
 
-import com.koddy.server.file.domain.model.RawFileData;
 import com.koddy.server.member.domain.model.Language;
 import com.koddy.server.member.domain.model.Nationality;
 import com.koddy.server.member.domain.model.mentor.Schedule;
@@ -12,7 +11,7 @@ public record CompleteMentorCommand(
         Long mentorId,
         String name,
         Nationality nationality,
-        RawFileData profile,
+        String profileUploadUrl,
         List<Language> languages,
         UniversityProfile universityProfile,
         String meetingUrl,
